+++ crates/models/src/log.rs
@@ -682,7 +682,7 @@ pub struct VmDirectWithdrawReport {
 #[derive(Copy, Clone, Zeroable, Pod, Default)]
 pub struct KaminoChangePositionReport {
     pub tag: u8,
-    pub padding_u8: u8,
+    pub assets_is_collateral: u8,
     pub padding_u16: u16,
     pub padding_u32: u32,
     pub seq_no: u32,