+++ crates/models/src/state/types.rs
@@ -113,7 +113,6 @@ pub mod account_type {
     pub const PERP_REBALANCE_TIME_TREE: u32 = 50;
     pub const PRIVATE_CLIENTS: u32 = 51;
     pub const VM_CLIENT: u32 = 52;
-    pub const KAMINO_CLIENT: u32 = 53;
 
     #[derive(Debug, Clone, Copy, PartialEq, Eq, Serialize, Deserialize)]
     #[repr(u32)]