+++ CHANGELOG.md
@@ -5,6 +5,12 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.1.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
+## [v2.58]
+
+## Added 
+- `kamino` flags
+- `max_withdraw` realtead `kamino` models
+
 ## [v2.57] - 2026-05-22
 
 ## Added